Rewrite example to use getStaticProps
